+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v ?.?.? - ?? ??? ????
 - OpenGLEngine tracks the state of the Caps Lock and Num Lock keys during key callbacks
+- OpenGL queries checks Point Size Range
 
 ## v 5.1.0 - 10 May 2024
 - Added CSCI441::drawSolidDome() and CSCI441::drawWireDome() which is a half-sphere with an open bottom